+++ packages/client-flutter/bin/reverse.dart
@@ -1,0 +1,3 @@
+import 'package:owlscope/src/reverse_runner.dart';
+
+void main() => runReverse();